fix test bugs introduced when merging to ts
