+++ test/2_endpoints_tests/1_createSession/createSession.js
@@ -4,7 +4,7 @@ const chaiHttp = require('chai-http');
 
 const { expect } = chai;
 const should = chai.should();
-const driver = require('../../../pluma-webdriver');
+const driver = require('../../../build/index');
 
 
 chai.use(chaiHttp);
