+++ test/2_endpoints_tests/6_elements/elements.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const chaiHttp = require('chai-http');
 
 const { expect } = chai;
-const driver = require('../../../pluma-webdriver');
+const driver = require('../../../build/index');
 
 chai.use(chaiHttp);
