+++ test/2_endpoints_tests/7_script/script.js
@@ -4,7 +4,7 @@ const chaiHttp = require('chai-http');
 
 const { expect } = chai;
 const should = chai.should();
-const driver = require('../../../pluma-webdriver');
+const driver = require('../../../build/index');
 
 chai.use(chaiHttp);
 
@@ -12,13 +12,8 @@ describe('Testing Cookies related enpoints\
     \n\t- Get all cookies GET /session/:sessionId/cookies\
     \n\t- Add Cookie POST /session/:sessionId/cookie', function () {
 
-    // TODO: test getting all cookies
-    describe('Get all cookies', function() {
+    describe('Test', function() {
         it('should do something here...');
     });
-
-    // TODO: test adding a valid cookie
-
-    // TODO: test adding invalid cookie
 
 });
